+++ Plugins/ArticyXImporter/Source/ArticyRuntime/Private/ArticyBaseObject.cpp
@@ -5,6 +5,7 @@
 #include "ArticyBaseObject.h"
 #include "ArticyPrimitive.h"
 #include "ArticyTypeSystem.h"
+#include "ArticyHelpers.h"
 
 UArticyPrimitive* UArticyBaseObject::GetSubobject(FArticyId Id) const
 {
@@ -20,4 +21,9 @@ void UArticyBaseObject::AddSubobject(UArticyPrimitive* Obj)
 FArticyType UArticyBaseObject::GetArticyType() const
 {
 	return ArticyType;
+}
+
+FText UArticyBaseObject::GetPropertyText(const FText Property)
+{
+	return ArticyHelpers::LocalizeString(this, Property, true, &Property);
 }